feat: rename payment pointer files
